requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.3
supabase==2.4.0
ably==1.2.0
python-dotenv==1.0.1